Rename graph.tsx => Graph.tsx
