+++ extensions/ql-vscode/src/view/results/Graph.tsx
